# perf: use fast property access instead of get were possible