--- a/lib/cast.js
+++ b/lib/cast.js
@@ -102,15 +102,18 @@ module.exports = function cast(schema, obj, options, context) {
           const pathFirstHalf = split.slice(0, j).join('.');
           const pathLastHalf = split.slice(j).join('.');
           const _schematype = schema.path(pathFirstHalf);
-          const discriminatorKey = get(_schematype, 'schema.options.discriminatorKey');
+          const discriminatorKey = _schematype &&
+            _schematype.schema &&
+            _schematype.schema.options &&
+            _schematype.schema.options.discriminatorKey;
 
           // gh-6027: if we haven't found the schematype but this path is
           // underneath an embedded discriminator and the embedded discriminator
           // key is in the query, use the embedded discriminator schema
           if (_schematype != null &&
-              get(_schematype, 'schema.discriminators') != null &&
-              discriminatorKey != null &&
-              pathLastHalf !== discriminatorKey) {
+            (_schematype.schema && _schematype.schema.discriminators) != null &&
+            discriminatorKey != null &&
+            pathLastHalf !== discriminatorKey) {
             const discriminatorVal = get(obj, pathFirstHalf + '.' + discriminatorKey);
             if (discriminatorVal != null) {
               schematype = _schematype.schema.discriminators[discriminatorVal].