--- a/lib/connection.js
+++ b/lib/connection.js
@@ -837,9 +837,19 @@ function _setClient(conn, client, options, dbName) {
   const db = dbName != null ? client.db(dbName) : client.db();
   conn.db = db;
   conn.client = client;
-  conn.host = get(client, 's.options.hosts.0.host', void 0);
-  conn.port = get(client, 's.options.hosts.0.port', void 0);
-  conn.name = dbName != null ? dbName : get(client, 's.options.dbName', void 0);
+  conn.host = client &&
+    client.s &&
+    client.s.options &&
+    client.s.options.hosts &&
+    client.s.options.hosts[0] &&
+    client.s.options.hosts[0].host || void 0;
+  conn.port = client &&
+    client.s &&
+    client.s.options &&
+    client.s.options.hosts &&
+    client.s.options.hosts[0] &&
+    client.s.options.hosts[0].port || void 0;
+  conn.name = dbName != null ? dbName : client && client.s && client.s.options && client.s.options.dbName || void 0;
   conn._closeCalled = client._closeCalled;
 
   const _handleReconnect = () => {
@@ -855,7 +865,10 @@ function _setClient(conn, client, options, dbName) {
     }
   };
 
-  const type = get(client, 'topology.description.type', '');
+  const type = client &&
+  client.topology &&
+  client.topology.description &&
+  client.topology.description.type || '';
 
   if (type === 'Single') {
     client.on('serverDescriptionChanged', ev => {