--- a/lib/document.js
+++ b/lib/document.js
@@ -1119,7 +1119,7 @@ Document.prototype.$set = function $set(path, val, type, options) {
 
     // `_skipMinimizeTopLevel` is because we may have deleted the top-level
     // nested key to ensure key order.
-    const _skipMinimizeTopLevel = get(options, '_skipMinimizeTopLevel', false);
+    const _skipMinimizeTopLevel = options._skipMinimizeTopLevel || false;
     if (len === 0 && _skipMinimizeTopLevel) {
       delete options._skipMinimizeTopLevel;
       if (val) {
@@ -2539,7 +2539,9 @@ function _getPathsToValidate(doc) {
         // To avoid potential performance issues, skip doc arrays whose children
         // are not required. `getPositionalPathType()` may be slow, so avoid
         // it unless we have a case of #6364
-        (!Array.isArray(_pathType) && _pathType.$isMongooseDocumentArray && !get(_pathType, 'schemaOptions.required'))) {
+        (!Array.isArray(_pathType) &&
+          _pathType.$isMongooseDocumentArray &&
+          !(_pathType && _pathType.schemaOptions && _pathType.schemaOptions.required))) {
       continue;
     }
 
@@ -2616,7 +2618,7 @@ Document.prototype.$__validate = function(pathsToValidate, options, callback) {
       (typeof options === 'object') &&
       ('validateModifiedOnly' in options);
 
-  const pathsToSkip = get(options, 'pathsToSkip', null);
+  const pathsToSkip = (options && options.pathsToSkip) || null;
 
   let shouldValidateModifiedOnly;
   if (hasValidateModifiedOnlyOption) {
@@ -3455,8 +3457,11 @@ Document.prototype.$toObject = function(options, json) {
   };
 
   const path = json ? 'toJSON' : 'toObject';
-  const baseOptions = get(this, 'constructor.base.options.' + path, {});
-  const schemaOptions = get(this, '$__schema.options', {});
+  const baseOptions = this.constructor &&
+  this.constructor.base &&
+  this.constructor.base.options &&
+  get(this.constructor.base.options, path) || {};
+  const schemaOptions = this.$__schema && this.$__schema.options || {};
   // merge base default options with Schema's set default options if available.
   // `clone` is necessary here because `utils.options` directly modifies the second input.
   defaultOptions = utils.options(defaultOptions, clone(baseOptions));
@@ -3503,7 +3508,7 @@ Document.prototype.$toObject = function(options, json) {
   }
 
   const depopulate = options.depopulate ||
-    get(options, '_parentOptions.depopulate', false);
+    (options._parentOptions && options._parentOptions.depopulate || false);
   // _isNested will only be true if this is not the top level document, we
   // should never depopulate
   if (depopulate && options._isNested && this.$__.wasPopulated) {
@@ -3780,7 +3785,9 @@ function applyVirtuals(self, json, options, toObjectOptions) {
   let assignPath;
   let cur = self._doc;
   let v;
-  const aliases = get(toObjectOptions, 'aliases', true);
+  const aliases = typeof (toObjectOptions && toObjectOptions.aliases) === 'boolean'
+    ? toObjectOptions.aliases
+    : true;
 
   let virtualsToApply = null;
   if (Array.isArray(options.virtuals)) {
@@ -4317,7 +4324,7 @@ Document.prototype.depopulate = function(path) {
 
   let populatedIds;
   const virtualKeys = this.$$populatedVirtuals ? Object.keys(this.$$populatedVirtuals) : [];
-  const populated = get(this, '$__.populated', {});
+  const populated = this.$__ && this.$__.populated || {};
 
   if (arguments.length === 0) {
     // Depopulate all