--- a/lib/drivers/node-mongodb-native/collection.js
+++ b/lib/drivers/node-mongodb-native/collection.js
@@ -8,7 +8,6 @@ const MongooseCollection = require('../../collection');
 const MongooseError = require('../../error/mongooseError');
 const Collection = require('mongodb').Collection;
 const ObjectId = require('./objectid');
-const get = require('../../helpers/get');
 const getConstructorName = require('../../helpers/getConstructorName');
 const stream = require('stream');
 const util = require('util');
@@ -76,7 +75,11 @@ function iter(i) {
     const collection = this.collection;
     const args = Array.from(arguments);
     const _this = this;
-    const debug = get(_this, 'conn.base.options.debug');
+    const debug = _this &&
+      _this.conn &&
+      _this.conn.base &&
+      _this.conn.base.options &&
+      _this.conn.base.options.debug;
     const lastArg = arguments[arguments.length - 1];
     const opId = new ObjectId();
 
@@ -84,7 +87,7 @@ function iter(i) {
     if (this.conn.$wasForceClosed) {
       const error = new MongooseError('Connection was force closed');
       if (args.length > 0 &&
-          typeof args[args.length - 1] === 'function') {
+        typeof args[args.length - 1] === 'function') {
         args[args.length - 1](error);
         return;
       } else {
@@ -368,7 +371,12 @@ function format(obj, sub, color, shell) {
           formatDate(x, key, shell);
         } else if (_constructorName === 'ClientSession') {
           x[key] = inspectable('ClientSession("' +
-            get(x[key], 'id.id.buffer', '').toString('hex') + '")');
+            (
+              x[key] &&
+              x[key].id &&
+              x[key].id.id &&
+              x[key].id.id.buffer || ''
+            ).toString('hex') + '")');
         } else if (Array.isArray(x[key])) {
           x[key] = x[key].map(map);
         } else if (error != null) {