--- a/lib/error/cast.js
+++ b/lib/error/cast.js
@@ -5,7 +5,6 @@
  */
 
 const MongooseError = require('./mongooseError');
-const get = require('../helpers/get');
 const util = require('util');
 
 /**
@@ -111,7 +110,9 @@ function getValueType(value) {
 }
 
 function getMessageFormat(schemaType) {
-  const messageFormat = get(schemaType, 'options.cast', null);
+  const messageFormat = schemaType &&
+  schemaType.options &&
+  schemaType.options.cast || null;
   if (typeof messageFormat === 'string') {
     return messageFormat;
   }