--- a/lib/index.js
+++ b/lib/index.js
@@ -19,7 +19,6 @@ const Query = require('./query');
 const Model = require('./model');
 const applyPlugins = require('./helpers/schema/applyPlugins');
 const driver = require('./driver');
-const get = require('./helpers/get');
 const promiseOrCallback = require('./helpers/promiseOrCallback');
 const legacyPluralize = require('./helpers/pluralize');
 const utils = require('./utils');
@@ -632,10 +631,8 @@ Mongoose.prototype._applyPlugins = function(schema, options) {
   const _mongoose = this instanceof Mongoose ? this : mongoose;
 
   options = options || {};
-  options.applyPluginsToDiscriminators = get(_mongoose,
-    'options.applyPluginsToDiscriminators', false);
-  options.applyPluginsToChildSchemas = get(_mongoose,
-    'options.applyPluginsToChildSchemas', true);
+  options.applyPluginsToDiscriminators = _mongoose.options && _mongoose.options.applyPluginsToDiscriminators || false;
+  options.applyPluginsToChildSchemas = typeof (_mongoose.options && _mongoose.options.applyPluginsToDiscriminators) === 'boolean' ? _mongoose.options.applyPluginsToDiscriminators : true;
   applyPlugins(schema, _mongoose.plugins, options, '$globalPluginsApplied');
 };