--- a/lib/model.js
+++ b/lib/model.js
@@ -710,7 +710,7 @@ Model.prototype.$__delta = function() {
   where._id = this._doc._id;
   // If `_id` is an object, need to depopulate, but also need to be careful
   // because `_id` can technically be null (see gh-6406)
-  if (get(where, '_id.$__', null) != null) {
+  if ((where && where._id && where._id.$__ || null) != null) {
     where._id = where._id.toObject({ transform: false, depopulate: true });
   }
   for (; d < len; ++d) {
@@ -1159,7 +1159,7 @@ Model.discriminator = function(name, schema, options) {
 
   options = options || {};
   const value = utils.isPOJO(options) ? options.value : options;
-  const clone = get(options, 'clone', true);
+  const clone = typeof options.clone === 'boolean' ? options.clone : true;
 
   _checkContext(this, 'discriminator');
 
@@ -1354,15 +1354,24 @@ Model.createCollection = function createCollection(options, callback) {
     options = void 0;
   }
 
-  const schemaCollation = get(this, ['schema', 'options', 'collation'], null);
+  const schemaCollation = this &&
+  this.schema &&
+  this.schema.options &&
+  this.schema.options.collation || null;
   if (schemaCollation != null) {
     options = Object.assign({ collation: schemaCollation }, options);
   }
-  const capped = get(this, ['schema', 'options', 'capped']);
+  const capped = this &&
+  this.schema &&
+  this.schema.options &&
+   this.schema.options.capped;
   if (capped) {
     options = Object.assign({ capped: true }, capped, options);
   }
-  const timeseries = get(this, ['schema', 'options', 'timeseries']);
+  const timeseries = this &&
+  this.schema &&
+  this.schema.options &&
+  this.schema.options.timeseries;
   if (timeseries != null) {
     options = Object.assign({ timeseries }, options);
   }
@@ -2588,7 +2597,7 @@ Model.findOneAndUpdate = function(conditions, update, options, callback) {
  */
 
 function _decorateUpdateWithVersionKey(update, options, versionKey) {
-  if (!versionKey || !get(options, 'upsert', false)) {
+  if (!versionKey || !(options && options.upsert || false)) {
     return;
   }
 
@@ -3327,10 +3336,10 @@ Model.$__insertMany = function(arr, options, callback) {
   }
   callback = callback || utils.noop;
   options = options || {};
-  const limit = get(options, 'limit', 1000);
-  const rawResult = get(options, 'rawResult', false);
-  const ordered = get(options, 'ordered', true);
-  const lean = get(options, 'lean', false);
+  const limit = options.limit || 1000;
+  const rawResult = !!options.rawResult;
+  const ordered = typeof options.ordered === 'boolean' ? options.ordered : true;
+  const lean = !!options.lean;
 
   if (!Array.isArray(arr)) {
     arr = [arr];
@@ -3408,12 +3417,12 @@ Model.$__insertMany = function(arr, options, callback) {
         // `writeErrors` is a property reported by the MongoDB driver,
         // just not if there's only 1 error.
         if (error.writeErrors == null &&
-            get(error, 'result.result.writeErrors') != null) {
+           (error.result && error.result.result && error.result.result.writeErrors) != null) {
           error.writeErrors = error.result.result.writeErrors;
         }
 
         // `insertedDocs` is a Mongoose-specific property
-        const erroredIndexes = new Set(get(error, 'writeErrors', []).map(err => err.index));
+        const erroredIndexes = new Set((error && error.writeErrors || []).map(err => err.index));
 
         let firstErroredIndex = -1;
         error.insertedDocs = docAttributes.
@@ -3634,7 +3643,7 @@ Model.bulkSave = function(documents, options) {
     .then(() => documents.map(buildSuccessfulWriteHandlerPromise))
     .then(() => bulkWriteResultPromise)
     .catch((err) => {
-      if (!get(err, 'writeErrors.length')) {
+      if (!(err && err.writeErrors && err.writeErrors.length)) {
         throw err;
       }
       return Promise.all(
@@ -3979,7 +3988,7 @@ Model.updateOne = function updateOne(conditions, doc, options, callback) {
 Model.replaceOne = function replaceOne(conditions, doc, options, callback) {
   _checkContext(this, 'replaceOne');
 
-  const versionKey = get(this, 'schema.options.versionKey', null);
+  const versionKey = this && this.schema && this.schema.options && this.schema.options.versionKey || null;
   if (versionKey && !doc[versionKey]) {
     doc[versionKey] = 0;
   }
@@ -4004,7 +4013,10 @@ function _update(model, op, conditions, doc, options, callback) {
   }
   options = typeof options === 'function' ? options : utils.clone(options);
 
-  const versionKey = get(model, 'schema.options.versionKey', null);
+  const versionKey = model &&
+  model.schema &&
+  model.schema.options &&
+  model.schema.options.versionKey || null;
   _decorateUpdateWithVersionKey(doc, options, versionKey);
 
   return mq[op](conditions, doc, options, callback);
@@ -4183,7 +4195,7 @@ Model.mapReduce = function mapReduce(o, callback) {
 Model.aggregate = function aggregate(pipeline, options, callback) {
   _checkContext(this, 'aggregate');
 
-  if (arguments.length > 3 || get(pipeline, 'constructor.name') === 'Object') {
+  if (arguments.length > 3 || (pipeline && pipeline.constructor && pipeline.constructor.name) === 'Object') {
     throw new MongooseError('Mongoose 5.x disallows passing a spread of operators ' +
       'to `Model.aggregate()`. Instead of ' +
       '`Model.aggregate({ $match }, { $skip })`, do ' +
@@ -4506,7 +4518,10 @@ function populate(model, docs, options, callback) {
     ids = utils.array.unique(ids);
 
     const assignmentOpts = {};
-    assignmentOpts.sort = get(mod, 'options.options.sort', void 0);
+    assignmentOpts.sort = mod &&
+      mod.options &&
+      mod.options.options &&
+      mod.options.options.sort || void 0;
     assignmentOpts.excludeId = excludeIdReg.test(select) || (select && select._id === 0);
 
     if (ids.length === 0 || ids.every(utils.isNullOrUndefined)) {
@@ -4675,7 +4690,9 @@ function _assign(model, vals, mod, assignmentOpts) {
   const isVirtual = mod.isVirtual;
   const justOne = mod.justOne;
   let _val;
-  const lean = get(options, 'options.lean', false);
+  const lean = options &&
+    options.options &&
+    options.options.lean || false;
   const len = vals.length;
   const rawOrder = {};
   const rawDocs = {};