--- a/lib/query.js
+++ b/lib/query.js
@@ -18,7 +18,6 @@ const castArrayFilters = require('./helpers/update/castArrayFilters');
 const castNumber = require('./cast/number');
 const castUpdate = require('./helpers/query/castUpdate');
 const completeMany = require('./helpers/query/completeMany');
-const get = require('./helpers/get');
 const promiseOrCallback = require('./helpers/promiseOrCallback');
 const getDiscriminatorByValue = require('./helpers/discriminator/getDiscriminatorByValue');
 const hasDollarKeys = require('./helpers/query/hasDollarKeys');
@@ -110,7 +109,10 @@ function Query(conditions, options, model, collection) {
   // versions of MongoDB
   this.$useProjection = true;
 
-  const collation = get(this, 'schema.options.collation', null);
+  const collation = this &&
+    this.schema &&
+    this.schema.options &&
+    this.schema.options.collation || null;
   if (collation != null) {
     this.options.collation = collation;
   }
@@ -1911,7 +1913,10 @@ Query.prototype._optionsForExec = function(model) {
   // Apply schema-level `writeConcern` option
   applyWriteConcern(model.schema, options);
 
-  const readPreference = get(model, 'schema.options.read');
+  const readPreference = model &&
+  model.schema &&
+  model.schema.options &&
+  model.schema.options.read;
   if (!('readPreference' in options) && readPreference) {
     options.readPreference = readPreference;
   }
@@ -2184,7 +2189,7 @@ Query.prototype._find = wrapThunk(function(callback) {
   // Separate options to pass down to `completeMany()` in case we need to
   // set a session on the document
   const completeManyOptions = Object.assign({}, {
-    session: get(this, 'options.session', null)
+    session: this && this.options && this.options.session || null
   });
 
   const cb = (err, docs) => {
@@ -3332,8 +3337,11 @@ Query.prototype.findOneAndUpdate = function(criteria, doc, options, callback) {
     delete options.fields;
   }
 
-
-  const returnOriginal = get(this, 'model.base.options.returnOriginal');
+  const returnOriginal = this &&
+    this.model &&
+    this.model.base &&
+    this.model.base.options &&
+    this.model.base.options.returnOriginal;
   if (options.new == null && options.returnDocument == null && options.returnOriginal == null && returnOriginal != null) {
     options.returnOriginal = returnOriginal;
   }
@@ -3661,7 +3669,11 @@ Query.prototype.findOneAndReplace = function(filter, replacement, options, callb
 
   options = options || {};
 
-  const returnOriginal = get(this, 'model.base.options.returnOriginal');
+  const returnOriginal = this &&
+  this.model &&
+  this.model.base &&
+  this.model.base.options &&
+  this.model.base.options.returnOriginal;
   if (options.new == null && options.returnDocument == null && options.returnOriginal == null && returnOriginal != null) {
     options.returnOriginal = returnOriginal;
   }
@@ -4566,19 +4578,19 @@ Query.prototype.orFail = function(err) {
       case 'update':
       case 'updateMany':
       case 'updateOne':
-        if (get(res, 'modifiedCount') === 0) {
+        if (res && res.modifiedCount === 0) {
           throw _orFailError(err, this);
         }
         break;
       case 'findOneAndDelete':
       case 'findOneAndRemove':
-        if (get(res, 'lastErrorObject.n') === 0) {
+        if ((res && res.lastErrorObject && res.lastErrorObject.n) === 0) {
           throw _orFailError(err, this);
         }
         break;
       case 'findOneAndUpdate':
       case 'findOneAndReplace':
-        if (get(res, 'lastErrorObject.updatedExisting') === false) {
+        if ((res && res.lastErrorObject && res.lastErrorObject.updatedExisting) === false) {
           throw _orFailError(err, this);
         }
         break;
@@ -4946,11 +4958,11 @@ Query.prototype.populate = function() {
     const readPref = this.options.readPreference;
 
     for (const populateOptions of res) {
-      if (readConcern != null && get(populateOptions, 'options.readConcern') == null) {
+      if (readConcern != null && (populateOptions && populateOptions.options && populateOptions.options.readConcern) == null) {
         populateOptions.options = populateOptions.options || {};
         populateOptions.options.readConcern = readConcern;
       }
-      if (readPref != null && get(populateOptions, 'options.readPreference') == null) {
+      if (readPref != null && (populateOptions && populateOptions.options && populateOptions.options.readPreference) == null) {
         populateOptions.options = populateOptions.options || {};
         populateOptions.options.readPreference = readPref;
       }
@@ -4962,7 +4974,7 @@ Query.prototype.populate = function() {
   if (opts.lean != null) {
     const lean = opts.lean;
     for (const populateOptions of res) {
-      if (get(populateOptions, 'options.lean') == null) {
+      if ((populateOptions && populateOptions.options && populateOptions.options.lean) == null) {
         populateOptions.options = populateOptions.options || {};
         populateOptions.options.lean = lean;
       }