--- a/lib/queryhelpers.js
+++ b/lib/queryhelpers.js
@@ -26,9 +26,9 @@ exports.preparePopulationOptions = function preparePopulationOptions(query, opti
 
   // lean options should trickle through all queries
   if (options.lean != null) {
-    pop.
-      filter(p => get(p, 'options.lean') == null).
-      forEach(makeLean(options.lean));
+    pop
+      .filter(p => (p && p.options && p.options.lean) == null)
+      .forEach(makeLean(options.lean));
   }
 
   pop.forEach(opts => {
@@ -53,12 +53,12 @@ exports.preparePopulationOptionsMQ = function preparePopulationOptionsMQ(query, 
 
   // lean options should trickle through all queries
   if (options.lean != null) {
-    pop.
-      filter(p => get(p, 'options.lean') == null).
-      forEach(makeLean(options.lean));
+    pop
+      .filter(p => (p && p.options && p.options.lean) == null)
+      .forEach(makeLean(options.lean));
   }
 
-  const session = get(query, 'options.session', null);
+  const session = query && query.options && query.options.session || null;
   if (session != null) {
     pop.forEach(path => {
       if (path.options == null) {
@@ -272,7 +272,7 @@ exports.applyPaths = function applyPaths(fields, schema) {
     // Special case: if user has included a parent path of a discriminator key,
     // don't explicitly project in the discriminator key because that will
     // project out everything else under the parent path
-    if (!exclude && get(type, 'options.$skipDiscriminatorCheck', false)) {
+    if (!exclude && (type && type.options && type.options.$skipDiscriminatorCheck || false)) {
       let cur = '';
       for (let i = 0; i < pieces.length; ++i) {
         cur += (cur.length === 0 ? '' : '.') + pieces[i];
@@ -319,7 +319,7 @@ exports.handleDeleteWriteOpResult = function handleDeleteWriteOpResult(callback)
       return callback(error);
     }
     const mongooseResult = Object.assign({}, res.result);
-    if (get(res, 'result.n', null) != null) {
+    if ((res && res.result && res.result.n || null) != null) {
       mongooseResult.deletedCount = res.result.n;
     }
     if (res.deletedCount != null) {