--- a/lib/schema.js
+++ b/lib/schema.js
@@ -425,7 +425,7 @@ Schema.prototype.pick = function(paths, options) {
 
 Schema.prototype.defaultOptions = function(options) {
   this._userProvidedOptions = options == null ? {} : utils.clone(options);
-  const baseOptions = get(this, 'base.options', {});
+  const baseOptions = this.base && this.base.options || {};
 
   const strict = 'strict' in baseOptions ? baseOptions.strict : true;
   options = utils.options({
@@ -519,7 +519,7 @@ Schema.prototype.add = function add(obj, prefix) {
     if (key === '_id' && val === false) {
       continue;
     }
-    if (val instanceof VirtualType || get(val, 'constructor.name', null) === 'VirtualType') {
+    if (val instanceof VirtualType || (val.constructor && val.constructor.name || null) === 'VirtualType') {
       this.virtual(val);
       continue;
     }