--- a/lib/schema/SubdocumentPath.js
+++ b/lib/schema/SubdocumentPath.js
@@ -13,7 +13,6 @@ const $exists = require('./operators/exists');
 const castToNumber = require('./operators/helpers').castToNumber;
 const discriminator = require('../helpers/model/discriminator');
 const geospatial = require('./operators/geospatial');
-const get = require('../helpers/get');
 const getConstructor = require('../helpers/discriminator/getConstructor');
 const handleIdOption = require('../helpers/schema/handleIdOption');
 const internalToObjectOptions = require('../options').internalToObjectOptions;
@@ -164,7 +163,7 @@ SubdocumentPath.prototype.cast = function(val, doc, init, priorVal, options) {
   let subdoc;
 
   // Only pull relevant selected paths and pull out the base path
-  const parentSelected = get(doc, '$__.selected', {});
+  const parentSelected = doc && doc.$__ && doc.$__.selected || {};
   const path = this.path;
   const selected = Object.keys(parentSelected).reduce((obj, key) => {
     if (key.startsWith(path + '.')) {
@@ -301,7 +300,9 @@ SubdocumentPath.prototype.doValidateSync = function(value, scope, options) {
 SubdocumentPath.prototype.discriminator = function(name, schema, options) {
   options = options || {};
   const value = utils.isPOJO(options) ? options.value : options;
-  const clone = get(options, 'clone', true);
+  const clone = typeof options.clone === 'boolean'
+    ? options.clone
+    : true;
 
   if (schema.instanceOfSchema && clone) {
     schema = schema.clone();