--- a/lib/schematype.js
+++ b/lib/schematype.js
@@ -8,7 +8,6 @@ const MongooseError = require('./error/index');
 const SchemaTypeOptions = require('./options/SchemaTypeOptions');
 const $exists = require('./schema/operators/exists');
 const $type = require('./schema/operators/type');
-const get = require('./helpers/get');
 const handleImmutable = require('./helpers/schematype/handleImmutable');
 const isAsyncFunction = require('./helpers/isAsyncFunction');
 const immediate = require('./helpers/immediate');
@@ -1026,7 +1025,7 @@ SchemaType.prototype.required = function(required, message) {
   this.isRequired = true;
 
   this.requiredValidator = function(v) {
-    const cachedRequired = get(this, '$__.cachedRequired');
+    const cachedRequired = this && this.$__ && this.$__.cachedRequired;
 
     // no validation when this path wasn't selected in the query.
     if (cachedRequired != null && !this.$__isSelected(_this.path) && !this[documentIsModified](_this.path)) {