… lib/schema/array.js
… lib/schema/documentarray.js
… lib/types/array/methods/index.js
… lib/types/map.js