With this PR I replaced the get-method with fast property access were it is possible. I know that it looks not so sexy like the get call, but it has performance improvements as we dont need to split the path, traverse an array and stuff.

before:
$ MONGOOSE_DEV=1 node benchmarks/clone.js 
Simple x 140,319 ops/sec ±1.35% (92 runs sampled)
AllSchema x 23,986 ops/sec ±1.26% (90 runs sampled)

MONGOOSE_DEV=1 node benchmarks/validate.js 
invalid x 22,726 ops/sec ±0.62% (95 runs sampled)
valid x 61,786 ops/sec ±0.71% (93 runs sampled)

after
$ MONGOOSE_DEV=1 node benchmarks/clone.js 
Simple x 196,277 ops/sec ±0.72% (92 runs sampled)
AllSchema x 28,769 ops/sec ±0.41% (94 runs sampled)

$ MONGOOSE_DEV=1 node benchmarks/validate.js 
invalid x 23,488 ops/sec ±2.98% (95 runs sampled)
valid x 69,983 ops/sec ±1.05% (92 runs sampled

So atleast cloning is about 25 % faster.

We need more benchmarks imho... but hey. its something